feat: Add a build action
This patch adds a build Action that is triggered for all pull requests and requires that the code
builds and passes all tests successfully before allowing a merge.

TODO:
This can benefit from some linting.
We could force format the code and fail if it changes.
Also we might want to generate test coverage here?
Spellcheck might be nice as well.